+++ CHANGELOG.md
@@ -77,6 +77,7 @@ data_source_fiat:
 - Conversion tool: added parser for BscScan explorer.
 - Conversion tool: specify the local currency of the "Value" headers.
 - Conversion tool: added parser for Exodus wallet.
+- Conversion tool: added parser for Zerion explorer.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
 - Binance parser: use filename to determine if deposits or withdrawals.